requirements.txt
pytest==8.4.2
selene==2.0.0rc9
allure-python-commons==2.15.0
allure-pytest==2.15.0
selenium==4.35.0
python-dotenv==1.2.1
curlify==3.0.0
jsonschema==4.25.1
requests==2.32.5
appium-python-client==5.2.4
pydantic==2.12.4
webdriver-manager==4.0.2
config==0.5.1